requirements.txt
streamlit
sentence-transformers
openai
bcrypt
PyPDF2
python-docx
pandas
numpy
scikit-learn
python-dotenv
